plasmacomponents3: Use "down" instead of "pressed" for buttons
"down" on AbstractButton is whether the button is "visually down".
Generally, it's synonymous to "pressed", except it's writable.

This allows to override the press behavior and either mimic a
"pressed down" state when it's not technically pressed, or disable
it for when a control doesn't actually react to a press.