Don't crash on invalid applets or coronas
configview assumes everywhere corona will be valid, so always have it
assigned when possible. it will cause a "broken" config dialog, but
should
not crash under any circumstance.

Another part of the fix will be to actually not show the config button
anywhere so that broken dialog can't even be triggered

BUG:472283