Fix PopupPlasmaWindow::effectivePopupDirection being the wrong way around
When a popup would open above a panel this would result in
`effectivePopupDirection` being `Qt::BottomEdge` and similarly for the
other edges. Due to this, the condition the code that applied resizing
would disable the incorrect edge (the one on the opposite of the edge on
the panel), breaking popup resizing.

BUG: 478764
CCBUG: 477880