Create a theme instance very early, fix shadows
Keep a theme instance as a memeber to create one as soon as possible,
as Theme creation will set KSvg to correctly fetch images form the Plasma Theme.
This makes sure elements are correct, both in the dialog surface and the shadows.

prior to this krunner often lacked correct shadows