Remove old focus workaround
There was a workaround for an old bug (372476) that fixed the focus
to the FullRepresentation item, but some applets wanted focus not on
the root fullrepresentation item but on some descendant.

For instance it breaks the calculator plasmoid keyboard input.
Reverting that workaround on Plasma6 the original bug doesn't happen
anymore.

BUG:481967