Better positioning of elements in the available screen region
Sometimes we have to place a rectangle within the available screen QRegion.
Previously, we simply tried to move right/down it according to the
available QRect. This MR makes the positioning algorithm not rely on the
available QRect at all, and only use QRegion data. A detailed explanation
of the new approach is written in a code comment.