ContainmentItem: set correct constraints when wallpaper is ready
Not all wallpaper plugins set their loading status to keep the splash
screen on, which means if the current wallpaper plugin is ready, there
is no isLoadingChanged signal, so the ui constraints will never be
updated. This adds a check to fix it.

BUG: 482267
FIXED-IN: 6.0.2


(cherry picked from commit 1b154b3eb2b7617067dfd05d961e53238458714e)