Do not activate containment when one of its applet is activated
Reasoning is that some applets use "activated" when they would to trigger
their action, e.g. the show desktop applet. If the containment (here, the
panel) also gets activated as a result, it will get mouse input and a blue
line will appear. Previously this was avoided by some hacky code that (a)
apparently was not working for some people and (b) broke auto-hide and
dodge windows after clicking on "Minimize all".

This will also allow us to delete the hack in plasma-desktop.

BUG: 483941
BUG: 479084
FIXED-IN: 6.0.3