Create Plasma version of ListSectionHeader
This allows us to use it in Plasma ListViews, the same we use the
Kirigami version in desktop-styled ListViews.

It's just like the Kirigami version, but uses the horizontal line SVG
from the active Plasma theme instead of a Kirigami.Separator.

I don't love the code duplication, but this is the best we can do until
we implement a way of swapping out pieces of the Kirigami implementations
of these components at runtime.