Applet: Drop mostly dead code to save config
Applet had a mechanism to save the applet state based on a timer. This
was called on applet construction and when a shortcut is saved.

A shortcut saving does not need to call save, the configuration is
already modified and applets have an existing mechansim to request a
deferred flush to disk.