Drop timetracker
The idea of TimeTracker was to dump significant events with timestamps
into a log file when built in debug mode and with an env var set.
Whilst the idea of TimeTracker was nice, it didn't deliver.

The output was custom JSON that was lacked any visualisation tool, and
the tracking all property changes for Applet objects isn't a very useful
for QML applets.

In my many years of having to profile plasma, this isn't a tool I ever
see people reaching for.

You can achieve similar results by setting a QT_MESSAGE_PATTERN with
timestamps. If some wanted to build something better, we would be better
starting anew with ftrace.