Add temporary faux api to allow popups to constrain popup placement
We want popups to not overlap other panels. We do not know these
locations from within libplasma window code.

Short term, this patch adds a property to allow client code to set the
availableScreenGeometry that should be used to constrain the window.

Mid-term will be addressed properly in Plasma 6.2 with a singleton that
tracks available screen geometry which is described at
https://invent.kde.org/plasma/libplasma/-/issues/22#note_943229 with
pending MRs, but it's too late for 6.1

Long-term (on wayland) kwin will position the window semantically itself
and this comes for free.

The code is not the prettiest, but there is a plan scheduled and it's a
good test of the underlying workflow we want without committing to API.


(cherry picked from commit a75b2503801b08ad4c01078b8ac5572208ca7e7e)