Init at componentComplete
Always init as soon the component is complete so the qml infrastructure is there
but unconditionally, so not anymore on show on scene.
It's not significantly slower as the "hard" thing that's done on demand on scene enter which is
the qml items creation is not done in init() anymore.
This has a minor binary incompatibility but it's not in frameworks anymore

this makes global shotcuts work also for hidden items in the systray

in conjunction with https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/4454