plasma/pluginloader: Repair loading plasmoid plugins on add from explorer
In some cases, PluginLoader::loadApplet can be passed a path instead of
a plugin name. 70fea2c4b0ac026a769d02532eed6c6f749b6d9c removed the
local metadata cache but also removed the logic to turn a path into a
plugin name. Restore the plugin name handling so we can load the right
plugin again when an applet gets added from the Widget Explorer.

Ultimately this restores behaviour that I am unsure is correct, I think
it would be better if either the function only accepted Plasmoid names
and the relevant calling code was changed, or we use the KPackage
metadata to get the plugin to load somehow.

BUG: 488592