Workaround to possible infinite reize loops
check against the following edge case:
The compact representation preferred size is bigger than both the switch
size, and the full representation preferred size.
this can cause in the panel (when is quite big) an infinite resize loop, because
the applet size is bigger than the switch size, then it switches to full
representaiton that has a smaller hint. this causes a resize that will make it
switch to compact representation, making it grow again and switch again

This problem happens with the timer applet which has those incorrect
size hints

BUG:488974