Spinbox: use more appropriate icons for decrease and increase buttons
Previously they were using the list-remove and list-add icons,
which was totally incorrect. It happened to work with Breeze icons
because they were plain minus and plus signs. ...Until recently, when
list-remove was changed to be an X, exposing the bug.

Instead, let's ask for dedicated spinbox-decrease and spinbox-increase
icons, which are being added to Breeze. If these don't exist, it will
fall back to the existing ones and nothing will change.

3rd-party icon themes might still exhibit the bug if their list-add or
list-remove icons look weird, but that's a pre-existing issue, and with
this, they'll at least have a path to fix it.

BUG: 491312
FIXED-IN: 6.1.4