Handle floating point position and size for PlasmoidHeading logic
It seems like the x position of the digital clock PlasmoidHeading is 2e-23 (I did not see why it assumed that value, maybe has to do with fractional scaling?); and, the size is also a fraction short of the window width. Just for sure, I add here some logic to add with non-integer position and sizes, since the rounded corners depend on it. 

BUG:492436
FIXED-IN: 6.2.0