Don't offer to create Icon widget when dropping .desktop files
When dragging-and-dropping apps from Kickoff/Task Manager/etc. to the
desktop, no fewer than four options can be presented, one of them being
"Add Icon". This one is a particularly poor choice because it becomes
a widget that looks like a normal file, but does not behave according
to the normal file semantics:

- It can't be drag-and-dropped or selected the same way
- It always opens on single-click
- It's harder to figure out how to remove
- It can overlap normal desktop icons

Fixing these issues is not really possible due to the nature of what an
Icon applet simply *is*; they are baked into its nature as a widget. As
such, Icon widgets are not very well suited for being on the desktop
most of the time, so let's remove the option from the drop mmenu.

CCBUG: 446031