AppletQuickItem: fix crash when applet is stopped by DBus service
When an applet is stopped by its DBus service, the applet will be
deleted immediately in the system tray before AppletQuickItem is deleted,
so accessing the applet pointer in dtor will cause a crash. However, reducing the
preload weight in AppletQuickItem::init() can achieve the same effect.

SENTRY: PLASMA-WORKSPACE-FJG
FIXED-IN: 6.2.1