containment: emit screen geometry change correctly
A containment reports the screen geometry. A containment has a
lastScreen member which is the index of the screen it should be on.

We used to track changes to the screen before a containment was assigned
to a screen, except the screen geometry got removed in
a0b8bb2a7bfa153b6e5923de019ac8b50f1be395.

reactToScreenChanged is called whenever Containment::screen changes this
is where a containment is bound to an physical screen. It's a better
hook for announcing geometry changes rather than having consumers make
this assumption.

CCBUG: 481736


(cherry picked from commit e568e18e192def09f60d9d626d55a455eaf71d19)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>