Fix of desktop applets transparency
Fix the transparency of desktop applets when the kwin contrast effect is
enabled on X11

when the contrast effect availability changes and/or compositing
available changes, event compress the setting of the appropriante
KSvg selectors (opaque, translucent...)

previously the contrast effect was not initialized correctly at plasma
start, but was read only on the effect enabled change, so in
order for it to pick the correct svg, the effect had to be
disabled and reenabled again while plasma was running

also, the current state only managed the X11 case but not Wayland.
This commit fixes the X11 part.

CCBUG:473289