Synchronize _q_showWithoutActivating with the panel takes focus property
When a window is shown, QtWayland will attempt to activate the window
unless it has a special type or the _q_showWithoutActivating property
is set to true.

If that property is not set, it can result in providing conflicting input
hints to the compositor, e.g. the compositor is told that the given
surface accepts no focus and yet it's asked to activate it.

CCBUG: 495537