AppletQuickItem: Better handle missing mainScript
SharedQmlEnginePrivate::beginExecute checks whether the source is empty,
which is a very reasonable thing to warn about.

However in that case it won't create a component for it, which is problematic
because AppletQuickItem will query the component for errors, causing a crash.
The code guards against the component being null, but inconsistenly.

Instead always create a component. If the source is empty the component will
hold an error and the UI will show it

SENTRY: PLASMA-WORKSPACE-14EM