ToolTip: Set popupType Window
New in Qt 6.8, lets us put tooltips into a separate window
so they can exceed the parent window's bounds.

This also corrects the use of "inset", which, when negative,
is designed for adding shadows around items. Previously, we
could just bleed outside the item's bounds. However, with it
being a dedicated window, it's important to do this properly,
so Qt can account for that in its popup window placement
and sizing.