Don't highlight disabled items on hover
In the past we didn't have to do this because we simply bound
`hoverEnabled` to `enabled`. But then we moved away from that to make
tooltips for disabled items appear on hover. Thus, we need to check
for enablement when deciding when to use the hover state for hovered
items.

BUG: 472811
FIXED-IN: 6.3.2