Center vertically/horizontally PopupPlasmaWindows when they do not fit boundaries
Previously, whenever a PopupPlasmaWindow would not fit the given boundary, it
would be left-aligned to the parent element. This was not intuitive and would
be a significant change to when the popup could fit the boundary. For a better
transition, we vertically/horizontally center align the popup to the given
boundary when it's not possible to fit completely within it.

BUG:500390