plasmaquick: don't use KWinInfo in Wayland dialogs
PlasmaCore.Dialog tests whether the dialog is spawning in a window
ignoring the task manager or in a window with type NET::Dock. The second
check requires KWinInfo, which is not available in Wayland, causing an
error message to be printed whenever a PlasmaCore.Dialog is opened,
such as in the task manager's GroupDialog.

Instead, only perform this test if we are in X11, avoiding the error
message.

BUG: 500394


(cherry picked from commit a8e4cba850f695db68d7cc18bfe3ef8b4d2f5993)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>