PC3/SpinBox: don't read implicitsize of loader before it's ready
Spinbox uses a Loader to conditionally load the editable or non-editable
version of the textinput background. The Loader coming in confuses the
implicit size assignment and leads to a binding loop for a moment before
settling on a proper value. This seems harmless, but causes a warning to
be printed each time a PC3 SpinBox is created.

Instead, just wait until the Loader is ready, and consider its implicit
size to be 0 until then.

BUG: 500338
FIXED-IN: 6.4.0