PC3/ScrollBar: consider policy before size for visible
For the binding that establishes whether the scrollbar should be
visible, the size is checked before the policy is considered. This means
that even if the scrollbar policy is always on or always off, the
binding will still be recomputed in response to size changes, only to
end up at the same value anyway.

Instead, test policy first, and only consider size if the policy is
AsNeeded. This is a marginal efficiency gain, but there doesn't seem to
be a reason to not save the tiny amount of extra work.