plasmaquick: use categorized logging
plasmaquick has a logging category defined in CMakeLists, but does not
consistently use it.

This change sets the logging category to give users more fine-grained
control over logging, and hide some debug messages by default. And while
we're at it, we switch to the qCDebug etc. macros as they're in
principle a bit faster if logging is off.