plasmaquick/containmentitem: remove desktoptoolbox support
The desktop toolbox was removed with the introduction of the zoomout
edit mode for Plasma 6.1, while the panel toolbox remains.
ContainmentItem still tries to load both of them, but fails for the
desktop and warns about it.

Instead, only load the toolboc for the panel, which simplifies the code
and avoids the warning.