System for nested containments
This allows for containments-within-containments, with the config file structure nested more deeply.

This makes not necessary anymore the hack that allowed the systray  and grouped plasmoid to work.

It used to be a systray applet that had as a child the main item of the systray containment, giving the illusion
of the systray containment within the panel containment.

Now migrate to the systray containment to be directly a panel's applet.

No code is needed to be changed in the systray containment, only some data in its json metadata