plasma/applet_p: don't connect nullptr action
When an action is added to an applet's contextualActions, a connection
is made to trigger the action's removal from the list when it is
destroyed. However, some complex list operations on the qml side may
insert nullptr temporarily, such as inserting actions at an arbitrary
point which grows the list by adding a nullptr element and then shuffles
the existing elements around to get everything in its right place. In
that case, the append method will try to connect to nullptr being
destroyed, which fails with an error.

Instead, only connect to the action's destroyed signal if it's not null.