Port org.kde.plasma.plasmoid to declarative type registration
This one is a bit tricky since it's intertwined with other libraries/modules.

Another problem is that using QML_EXTENDED with Plasma::Applet doesn't seem to
work since Applet is not default constructible. We already get the properties
from Applet into PlasmoidAttached by being an attached property, we are only
missing the enums. To work around that copy the enum definition