plasmaquick: remove includeMargins property
This commit reverts 4910e2c4.

The includeMargins property was necessary with the old config page
infrastructure to allow Views to touch the edge of the page. In Plasma
6, all (non-containment) config pages have to be based on KCMUtils
components, where this is not necessary (and until recently was not even
available to these pages at all, so couldn't be used even if it was set
in config.qml). Containment config pages continued to use the old method
up to Plasma 6.3, but with f4f98b04 and ffcae8c2 these were switched to
using the new method as well. While the property is currently still
supported there, it's no longer necessary and it's simpler to set the
margins directly in the config page, as you would in kcms or app config
dialogs.

This changes removes support for this property from libplasma.