plasmaquick: remove loading kcms from configmodel
ConfigModel has the ability to load kcms if they are set as the
pluginName. This was used in the past, but is no longer - applets either
just use the kcm (and open it in System Settings) or have a button to
open System Settings on the relevant page. So we have no regular testing
for this feature. The ability to do this is also rather limited,
allowing only qml kcms. This not only adds complexity in the backend and
frontend, it's also causing issues for other uses of ConfigModel (such
as wallpaper plugins, which are also stored as a ConfigModel, and in
certain circumstances may attempt to be loaded as kcms instead because
ConfigModel has this functionality).

Instead, we just remove this unused functionality.