Rework how WallpaperItem loads its config
Currently wallpaperItem loads the config in classBegin.

In the screenlocker we are also using WallpaperItem, but there we
want to load a different config.

Currently this works by the screenlocker registering a different implementation under
the same name. This is rather hacky and falls apart when doing declarative type registration.

To avoid this make passing in the config the reponsibility of whoever is loading the item.