plasma/pluginloader: actually use pluginName when loading plugins
loadApplet takes the plugin name either as an applet name or as a full
path, but KPluginMetaData only wants the actual name. So we save the
plugin name in pluginName, then ignore it and use the original argument
anyway. WidgetExplorer, when clicking the applet, sends the path, so all
Kpackage + plugin widgets added this way are considered qml-only plugins
and their applet code isn't started (unless they have a parent plugin,
as the rootpath metadata is the name, or Plasma is restarted and the
name is read from appletsrc).

Instead we actually use the variable we computed and everything works
again.

BUG: 504343