Fix crash in drop job
When a mimedata is dragged onto the desktop libplasma goes through one
of two paths, one immediate, if multiple are present we show a menu and
act when one is selected.

The attached mimedata from a dropevent is only valid during the initial
handling. It is not safe to capture in a lambda.

This patch catpures data ahead of time.

This also fixes a bug that this async path could never have worked, it
needs to call into pluginFormats. Otherwise we're looking up the
mimedata 'org.kde.mediaFrame' rather than 'image/png'.

BUG: 504355