Destroy KWindowShadow objects in a post routine
Some windows (mainly panel popups) survive past the moment the QPA is
destroyed. This creates problems with wayland client extensions.

When the wl_display object is destroyed, the associated proxy objects
will be destroyed too. In other words, we are not allowed to use
wayland calls after the QGuiApplication destructor runs. It also implies
that you need to be very very careful if you create an object that uses
Wayland APIs and is a child of the application object.

In order to harden the corresponding logic, this change makes sure that
the managed shadow and shadow tile objects are destroyed before the QPA
is destroyed.

The API of the DialogShadows was changed to ensure that no new window
shadow objects are created by accident when plasma shuts down.

CCBUG: 487660