plasmaextracomponents: fix icon-less PlaceholderMessage
PlaceholderMessage sets the Kirigami.icon source to null if the iconName
is falsy, but then tests to see whether it's undefined to determine
visibility. Null is not undefined, so it's always visible (and takes up
space) even if unset.

Instead, set the source to undefined if iconName is falsy (similar to
what the Kirigami version does).


(cherry picked from commit ad838357530e6c7087246a262cb810577282cf8e)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>