ConfigView: Resize to implicitSize before saved size
Otherwise we will always overwrite the saved value with implicit size.
This way the implicit size serves as a default and in all other
cases we use the saved size. Also move the resize to mainItemLoaded
so the logic is in one place.


(cherry picked from commit 673d617fd269093c0472478c046a2ab8721f225a)

Co-authored-by: David Redondo <kde@david-redondo.de>