Fix compile without libxcb-composite
If libxcb-composite is not installed but GLX or EGL are
then compile fails in windowthumbnail.cpp.

The whole file makes use of the optional xcb_composite,
glx and egl code by using that logic:

  #if HAVE_XCB_COMPOSITE
  #if HAVE_GLX
    [glx code]
  #endif
  #if HAVE_EGL
    [egl code]
  #endif
  #endif

But in WindowThumbnail::invalidateSceneGraph() the
outer HAVE_XCB_COMPOSITE is missing resulting in
a compile error because inline classes like
DiscardGlxPixmapRunnable and DiscardEglPixmapRunnable
are not defined if HAVE_XCB_COMPOSITE is not the
case.

Fix that by adding the missing HAVE_XCB_COMPOSITE
condition also to WindowThumbnail::invalidateSceneGraph().