popupplasmawindow: Check if resizeEvents have resized anything
QtBase ensures a resize event is always sent before the first expose
event. Either on the expose itself or when application code triggers a
move.

If the size is unchanged, there is no need to reposition anything.

This also avoids a bug where a positional updatePosition on wayland can
loop if QT_WAYLAND_DISABLE_FIXED_POSITIONS is not set.

BUG: 506106