Theme: Listen to org.kde.KGlobalSettings notifyChange
a color from the plasma theme comes from KColorScheme frame contrast
property. this is not included in QPalette and when this changes, it
won't cause a PaletteChange event.

Listening to /KGlobalSettings org.kde.KGlobalSettings on the
notifyChange event is the only way to actually know when this, or any
other color not included is qpalette has actually changed